refactor: project layout
